feat: 크롤링 배치 기능 구현 및 실행 로직 개선
ModeMan 사이트의 상품 정보를 크롤링하는 Spring Batch 잡의 전체 흐름을 구현하고 기존 로직을 개선합니다.

주요 변경 사항:
- 배치 실행 로직 개선 (`ModeManBatchRunner`):
  - 커맨드 라인 인자(`key=value`)를 동적으로 Job 파라미터로 전달하도록 수정하여, 외부에서 크롤링 대상을 지정할 수 있게 합니다.

- 스텝 간 데이터 전달 (`ProductBrief`):
  - `ProductBrief` DTO에 `Serializable`을 구현하여, 목록 스텝에서 상세 스텝으로 수집 데이터를 전달할 때 발생하는 직렬화 오류를 해결합니다.

- JSON 출력 포맷 정리:
  - `ProductSnapshot`, `ProductOption` DTO에 `@JsonIgnore`를 추가하여 최종 NDJSON 결과물에 불필요한 내부 필드가 포함되지 않도록 합니다.

- 문서 추가 (`docs.md`):
  - 배치 실행 방법, 아키텍처, 주요 파일의 역할 등을 상세히 기술한 문서를 추가하여 유지보수성을 향상시킵니다.

- 기타 개선:
  - 배치 설정을 2-step (목록 -> 상세) 구조로 완성합니다.
  - `Tasklet`과 `Processor`가 Job 파라미터를 받을 수 있도록 `@StepScope`를 적용합니다.
  - 전반적인 DTO 및 파서 로직을 개선하고 안정화합니다.

+++ src/main/java/com/jj/redline/domain/dto/CategoryDto.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CategoryDto {
-    private int code;      // cate_no (예: 263, 858)
+    private long code;      // [수정] cate_no (예: 263, 858)
     private String name;   // "Denim Jackets", "Denim Pents"
 }
